back/requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
# actualizar a Django 5 en 2024
Django==4.1.6
django-environ==0.9.0
django-filter==2.4.0
djangorestframework==3.13.1
kombu==5.2.4
Markdown==3.3.6
prompt-toolkit==3.0.37
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.6
psycopg2==2.9.5
redis==4.5.1
celery==5.2.7
importlib-metadata==4.13.0
# para recibir emails
imaplib2==3.6
python-dotenv==1.0.0
django-celery-beat==2.4.0
django_celery_results==2.4.0

